+++ wls-gui-wahllokalsystem/src/components/wahlvorbereitung/BaseWahleroeffnungCard.vue
